fix: move some more files
