+++ libs/sdk/package.json
@@ -14,6 +14,7 @@
   "devDependencies": {
     "@eventuras/typescript-config": "*",
     "node-fetch": "^3.3.2",
+    "openapi-typescript-codegen": "^0.28.0",
     "typescript": "^5.4.3",
     "vite": "^5.2.7",
     "vite-plugin-dts": "^3.8.1"
