+++ libs/ui/src/index.ts
@@ -1,8 +1,9 @@
 import './styles/globals.css';
-export { default as AppNotifications } from './AppNotifications';
-export type { AppNotificationOptions, AppNotificationType } from './AppNotifications';
+export { default as AppNotifications, AppNotificationType } from './AppNotifications';
+export type { AppNotificationOptions } from './AppNotifications';
 export { default as Badge } from './Badge';
 export { default as Button, buttonStyles } from './Button';
+export { default as ButtonGroup } from './ButtonGroup';
 export { spacingClassName } from './Box';
 export type { BoxProps } from './Box';
 export { default as Container } from './Container';
@@ -11,8 +12,8 @@ export { default as Drawer } from './Drawer';
 export { default as Grid } from './Grid';
 export { default as Heading } from './Heading';
 export { default as Loading } from './Loading';
+export { default as NumberCard } from './NumberCard';
 export { default as Pagination } from './Pagination';
 export { default as Portal } from './Portal';
 export { default as Section } from './Section';
 export { default as Text } from './Text';
-export { default as InputAutoComplete } from './InputAutoComplete'
